course::handlers: add, remove section implementation
Implements the case where course has an implicit link to all sections
included into it. This scheme allows for fast reads of all the
sections in the course but makes for slower writes (when adding /
removing section from the course). It makes sense given the design of an
app, where multiple students would be performing way more reads than
amount of writes performed by a single teacher.