course, section: accept timestamp on updates
This fixes an update loop bug where adding and then removing section
from the course results in a timeout. Bug is reproduced as:
1. create an empty course: would result in some course entry, let's
callt it course_v1;
2. create a section inside this course: would result in another course
entry with a section, let's call it course_v2;
3. remove a section from this course: would result in an empty course
again that would be equal to course_v1 (because timestamp isn't
updated). And we'll be failing with a timeout from the zome.

Problem here is that by the step 3 link from CourseAnchor to course_v1
has already been deleted, but we're trying to create it once again
(thinking that step 3 produces a new, unique course entry).

Now that we have timestamp value, step 3 would result in a different
entry course_v3 and link creation from CourseAnchor to course_v3 would
work as expected.

Timestamps are also accepted for the section to keep all it's entries
unique as well.